[MOD] 토스트메시지 및 네트워킹 알림 수정
